deployment/sandbox-v2/utils/reprint/requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
importlib-metadata==3.3.0
iniconfig==1.1.1
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dotenv==0.12.0
requests==2.25.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
zipp==3.4.0
